Разнес модуль gameplay по файлам
