+++ js/gameplay/gameplay.getTimeLeft.test.js
@@ -1,5 +1,5 @@
 import {assert} from 'chai';
-import gameplay from './gameplay';
+import getTimeLeft from './gameplay.getTimeLeft';
 
 describe(`Функция проверки таймера`, () => {
   let startDate;
@@ -9,19 +9,19 @@ describe(`Функция проверки таймера`, () => {
     checkDate = new Date(2018, 1, 1, 1, 1, 0, 1);
     startDate = new Date(2018, 1, 1, 1, 1, 15, 1);
     timer = 30;
-    assert.equal(gameplay.getTimeLeft(startDate, timer, checkDate), 15);
+    assert.equal(getTimeLeft(startDate, timer, checkDate), 15);
   });
   it(`возвращает 15 (секунд), при таймере в 30 с началом 15 секунд назад
       от текущего времени (без передачи аргумента проверяемой даты)`, () => {
     startDate = new Date();
     startDate.setTime(startDate.getTime() + 15000);
     timer = 30;
-    assert.equal(Math.floor(gameplay.getTimeLeft(startDate, timer)), 15);
+    assert.equal(Math.floor(getTimeLeft(startDate, timer)), 15);
   });
   it(`возвращает -1, если разница дат больше таймера и время вышло`, () => {
     startDate = new Date();
     startDate.setTime(startDate.getTime() + 32000);
     timer = 30;
-    assert.equal(Math.floor(gameplay.getTimeLeft(startDate, timer)), -1);
+    assert.equal(Math.floor(getTimeLeft(startDate, timer)), -1);
   });
 });
